ts: Add translate-shell.vim
This is simple wrapper for `translate-shell` application that does the
translation.  Translating easily directly from Vim is amazing and that
is why we are here.

As for now only mappings are used to invoke translation. Internal
mappings marked with <Plug> are defined and used in public mappings that
can be enabled using one of available global variables.

There are three ways you can translate text:

- translate a word under the cursor and echo it
- translate selection and echo it
- translate selection and write it to the temporary file that is shown
  in a new window

Those <Plug> mappings start translation process with some default
parameters.  Those parameters can easily be updated to one's needs.

This is first looking good version of this plugin. More to come.